src/fundings/requirements.txt
celery[redis]
psycopg2-binary
python-dotenv
asyncio
